chore: move to TelemetryProvider
